pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-label-factory"
version = "0.2.0"
description = "Generic auto-labeling pipeline for vision datasets — runs on a 16 GB Apple Silicon Mac via SSD-streaming MoE."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "walter-grace" },
]
keywords = [
    "vision",
    "dataset",
    "labeling",
    "annotation",
    "object-detection",
    "mlx",
    "apple-silicon",
    "qwen",
    "gemma",
    "falcon-perception",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "pyyaml>=6.0",
    "pillow>=9.0",
    "requests>=2.28",
]

[project.optional-dependencies]
gather = [
    # Image-search gathering (DDG, Wikimedia, YouTube frame extraction)
    "duckduckgo-search>=4.0",
    "yt-dlp>=2024.0.0",
]
runpod = [
    # Optional GPU path: orchestrate a RunPod pod, publish to HF when done
    "runpod>=1.7.0",
    "huggingface_hub>=0.26",
    "datasets>=3.0",
    "pyarrow>=17.0",
]
identify = [
    # Open-set CLIP retrieval: build/verify/train/serve a card-style index
    "torch>=2.1",
    "torchvision>=0.16",
    "numpy>=1.24,<2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.20",
    "pillow>=10.0",
    "ultralytics>=8.3",
    "clip @ git+https://github.com/openai/CLIP.git",
]
chandra = [
    # Chandra OCR 2 — document/text labeling with spatial annotations
    "chandra-ocr[hf]",
    "transformers>=4.40",
]
wilddet3d = [
    # WildDet3D — open-vocab 2D+3D detection (CUDA GPU required)
    "torch>=2.5",
    "torchvision>=0.20",
    "einops",
    "timm",
    "transformers>=4.40",
]
mcp = [
    # MCP server for AI agent integration
    "mcp>=1.0",
]
flywheel = [
    # Synthetic data generation (flywheel pattern)
    "opencv-python>=4.8",
    "numpy>=1.24",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/walter-grace/data-label-factory"
Repository = "https://github.com/walter-grace/data-label-factory"
HuggingFace = "https://huggingface.co/waltgrace/data-label-factory"
Issues = "https://github.com/walter-grace/data-label-factory/issues"

[project.scripts]
data_label_factory = "data_label_factory.cli:main"
data-label-factory = "data_label_factory.cli:main"

[tool.setuptools]
packages = [
    "data_label_factory",
    "data_label_factory.providers",
    "data_label_factory.runpod",
    "data_label_factory.identify",
]

[tool.setuptools.package-data]
data_label_factory = ["*.py"]
"data_label_factory.providers" = ["*.py"]
"data_label_factory.runpod" = ["*.py", "*.md", "Dockerfile", "requirements-pod.txt"]
"data_label_factory.identify" = ["*.py", "*.md"]
